Chyba poprawa układu plików
TO DO:
- błędy językowe
- przyciski
- kursy od trenera
- testy jednostkowe
- dokumentacja
