+++ BackEnd/projekt/src/main/resources/templates/Auth/login.html
@@ -12,7 +12,7 @@
 
 <nav>
     <ul>
-        <li><a th:href="@{/}">Home</a></li>
+        <li><a th:href="@{/static}">Home</a></li>
         <li><a th:href="@{/register}">Register</a></li>
         <li><a th:href="@{/login}">Login</a></li>
     </ul>
