+++ BackEnd/projekt/src/main/resources/templates/Course/Trainer/trainer-courses.html
