+++ BackEnd/projekt/src/main/resources/templates/Course/Users/all-courses.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Wszystkie kursy</title>
+    <link rel="stylesheet" th:href="@{/css/styles.css}">
+</head>
+<body>
+
+<header>
+    <h1>Wszystkie kursy</h1>
+</header>
+
+<!-- Base navigation fragment -->
+<div th:replace="base :: navigation"></div> 
+
+<div class="container">
+    <div class="card">
+        <h2>Lista Kursów</h2>
+
+        <div th:if="${courses.size() > 0}">
+            <ul>
+                <li th:each="course : ${courses}">
+                    <p>
+                        <strong>Tytuł:</strong> 
+                        <span th:text="${course.title}">Tytuł kursu</span><br>
+
+                        <strong>Liczba użytkowników zapisanych:</strong> 
+                        <span th:text="${course.countUsers}">0</span><br>
+
+                        <strong>Trener:</strong>
+                        <span th:text="${course.courseTrainer.firstName}">Imię</span>
+                        <span th:text="${course.courseTrainer.lastName}">Nazwisko</span>
+                        (<span th:text="${course.courseTrainer.specialization}">Specjalizacja</span>)
+                    </p>
+
+                    <!-- Action Buttons -->
+                    <div th:if="${user.getTrainer() == null || !user.getTrainer().getId().equals(course.getCourseTrainer().getId())}">
+                        <div class="actions" th:if="${!user.isUserExistInCourse(course.getId())}">
+                            <!-- Dołącz do kursu (POST) -->
+                            <form th:action="@{/course/user/{id}(id=${course.id})}" method="post" style="display:inline">
+                                <button type="submit" class="btn">Dołącz do kursu</button>
+                            </form>
+                        </div>
+
+                        <div class="actions" th:if="${user.isUserExistInCourse(course.getId())}">
+                            <form method="post" th:action="@{/course/user/{id}(id=${course.id})}">
+                                <input type="hidden" name="_method" value="delete"/>
+                                <button type="submit" class="btn">Opuść kurs</button>
+                            </form>
+                        </div>
+                    </div>
+
+                    <hr>
+                </li>
+            </ul>
+        </div>
+
+        <div th:if="${courses.size() == 0}">
+            <p>Nie masz jeszcze żadnych kursów.</p>
+        </div>
+    </div>
+</div>
+
+<footer>
+    <p>&copy; 2025 Twoja Aplikacja Treningowa</p>
+</footer>
+
+</body>
+</html>
