+++ BackEnd/projekt/src/main/resources/templates/Course/Users/my-courses.html
@@ -2,28 +2,28 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>Wszystkie kursy</title>
+    <title>Moje Kursy</title>
     <link rel="stylesheet" th:href="@{/css/styles.css}">
 </head>
 <body>
 
 <header>
-    <h1>Wszystkie kursy</h1>
+    <h1>Moje Kursy</h1>
 </header>
 
-<!-- Base navigation fragment -->
+<!-- Include base navigation -->
 <div th:replace="base :: navigation"></div> 
 
 <div class="container">
     <div class="card">
-        <h2>Lista Kursów</h2>
+        <h2>Lista Kursów, do których jesteś zapisany</h2>
 
-        <div th:if="${courses.size() > 0}">
+        <div th:if="${myCourses.size() > 0}">
             <ul>
-                <li th:each="course : ${courses}">
+                <li th:each="course : ${myCourses}">
                     <p>
                         <strong>Tytuł:</strong> 
-                        <span th:text="${course.title}">Tytuł kursu</span><br>
+                        <a th:href="@{/course/{id}(id=${course.id})}" th:text="${course.title}">Tytuł kursu</a><br>
 
                         <strong>Liczba użytkowników zapisanych:</strong> 
                         <span th:text="${course.countUsers}">0</span><br>
@@ -33,30 +33,18 @@
                         <span th:text="${course.courseTrainer.lastName}">Nazwisko</span>
                         (<span th:text="${course.courseTrainer.specialization}">Specjalizacja</span>)
                     </p>
-
-                    <!-- Action Buttons -->
-                    <div class="actions">
-                    <!-- Dołącz do kursu (POST) -->
-                    <form th:action="@{/course/user/{id}(id=${course.id})}" method="post" style="display:inline">
-                        <button type="submit" class="btn">Dołącz do kursu</button>
-                    </form>
-
                     <!-- Opuść kurs (DELETE) -->
                     <form method="post" th:action="@{/course/user/{id}(id=${course.id})}">
                     <input type="hidden" name="_method" value="delete"/>
-                        <button type="submit">Opuść kurs</button>
+                        <button type="submit" class="btn">Opuść kurs</button>
                     </form>
-
-
-                    </div>
-
                     <hr>
                 </li>
             </ul>
         </div>
 
-        <div th:if="${courses.size() == 0}">
-            <p>Nie masz jeszcze żadnych kursów.</p>
+        <div th:if="${myCourses.size() == 0}">
+            <p>Nie jesteś jeszcze zapisany na żaden kurs.</p>
         </div>
     </div>
 </div>
