+++ BackEnd/projekt/src/main/resources/templates/TrainingPlan/Users/user-training-plan-details.html
@@ -44,7 +44,7 @@
         </div>
 
         <!-- Back to plans list -->
-        <a th:href="@{/trainingplan}">
+        <a th:href="@{/trainingplan/user}">
             <button>Wróć do planów</button>
         </a>
