+++ BackEnd/projekt/src/main/resources/templates/WorkOut/User/user-workout-details.html
@@ -35,7 +35,7 @@
         </div>
 
         <!-- Back to workouts button -->
-        <a th:href="@{/users/workout}">
+        <a th:href="@{/workout/user}">
             <button>Wróć do listy treningów</button>
         </a>
