+++ BackEnd/projekt/src/main/resources/templates/base.html
@@ -17,25 +17,21 @@
         <!-- First row of navigation -->
         <ul>
             <li><a th:href="@{/home}">Strona główna</a></li>
-            <li><a th:href="@{/performLogout}">Wyloguj</a></li>
-        </ul>
-        <ul>
-            <li><a th:href="@{/course/all}">Wszystkie kursy</a></li>
+            <li><a th:href="@{/me}">Moje dane</a></li>
+            <li><a th:href="@{/course}">Wszystkie kursy</a></li>
             <li><a th:href="@{/course/user}">Moje kursy</a></li>
-            <li><a th:href="@{/course/trainer}" th:if="${user.trainer != null}">Kursy trenera</a></li>
-            <li><a th:href="@{/course/form}" th:if="${user.trainer != null}">Dodaj kurs</a></li>
+            <li><a th:href="@{/workout/user}">Moje treningi</a></li>
+            <li><a th:href="@{/trainingplan/user}">Moje plany treningowe</a></li>
         </ul>
 
-        <!-- Second row of navigation (new buttons) -->
-         
         <ul>
-            <li><a th:href="@{/users/me}">Moje dane</a></li>
-            <li><a th:href="@{/users/workout}">Moje treningi</a></li>
-            <li><a th:href="@{/users/trainingplan}">Moje plany treningowe</a></li>
-            <li><a th:href="@{/users/trainers}">Lista trenerów</a></li>
-            <li><a th:href="@{/users/workout/form}"> Utwórz trening</a></li>
-            <li><a th:href="@{/users/trainer}" th:if="${user.trainer == null}">Zostań trenerem</a></li>
-            
+            <li><a th:href="@{/trainer}">Lista trenerów</a></li>
+            <li><a th:href="@{/trainingplan/user/form}">Stwórz plan treningowy</a></li>
+            <li><a th:href="@{/course/trainer}" th:if="${user.trainer != null}">Kursy trenera</a></li>
+            <li><a th:href="@{/course/trainer/form}" th:if="${user.trainer != null}">Dodaj kurs</a></li>
+            <li><a th:href="@{/workout/user/form}"> Utwórz trening</a></li>
+            <li><a th:href="@{/trainer/form}" th:if="${user.trainer == null}">Zostań trenerem</a></li>
+            <li><a th:href="@{/performLogout}">Wyloguj</a></li>
         </ul>
         
     </nav>
